# rapla-app/src/main/resources/openapi/client.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Rapla REST API",
    "description" : "Resource scheduling and event planning. Authorize via OAuth2 (Authorization Code + PKCE) using Rapla's bundled Spring Authorization Server, or paste a raw bearer JWT.",
    "version" : "2.1-SNAPSHOT"
  },
  "security" : [ {
    "oauth2" : [ ]
  }, {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/storage/refresh" : {
      "post" : {
        "tags" : [ "remote-storage-controller" ],
        "operationId" : "refresh",
        "parameters" : [ {
          "name" : "lastValidated",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UpdateEvent"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/impersonate" : {
      "post" : {
        "tags" : [ "impersonation-controller" ],
        "operationId" : "impersonate",
        "parameters" : [ {
          "name" : "target_username",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "403" : {
            "description" : "Forbidden"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImpersonationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/impersonate/switch" : {
      "post" : {
        "tags" : [ "auth-cookie-controller" ],
        "operationId" : "impersonateSwitch",
        "parameters" : [ {
          "name" : "target_username",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/storage/user" : {
      "get" : {
        "tags" : [ "remote-storage-controller" ],
        "operationId" : "getUsername",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RaplaConfiguration" : {
        "properties" : {
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "children" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DefaultConfiguration"
            }
          },
          "value" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "AllocatableImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "classification" : {
            "$ref" : "#/components/schemas/ClassificationImpl"
          },
          "permissions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PermissionImpl"
            }
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "annotations" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
...
      "AttributeImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "$ref" : "#/components/schemas/MultiLanguageName"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "string", "int", "date", "boolean", "rapla:category", "rapla:allocatable" ]
          },
          "key" : {
            "type" : "string"
          },
          "multiSelect" : {
            "type" : "boolean"
          },
          "belongsTo" : {
            "type" : "boolean"
          },
          "packages" : {
            "type" : "boolean"
          },
          "optional" : {
            "type" : "boolean"
          },
          "annotations" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "defaultValue" : {
            "type" : "string"
          }
        }
      },
      "CalendarModelConfigurationImpl" : {
        "properties" : {
          "classificationFilters" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ClassificationFilterImpl"
            }
          },
          "selected" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "typeList" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "title" : {
            "type" : "string"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "selectedDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "view" : {
            "type" : "string"
          },
          "optionMap" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "defaultEventTypes" : {
            "type" : "boolean"
          },
          "defaultResourceTypes" : {
            "type" : "boolean"
          },
          "resourceRootSelected" : {
            "type" : "boolean"
          }
        }
      },
      "CategoryImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "$ref" : "#/components/schemas/MultiLanguageName"
          },
          "key" : {
            "type" : "string"
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "annotations" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
...
      "ClassificationImpl" : {
        "properties" : {
          "typeId" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        }
      },
      "ConflictImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reservation1Name" : {
            "type" : "string"
          },
          "reservation2Name" : {
            "type" : "string"
          },
          "repeatingType1" : {
            "type" : "string",
            "enum" : [ "weekly", "daily", "monthly", "yearly" ]
          },
          "repeatingType2" : {
            "type" : "string",
            "enum" : [ "weekly", "daily", "monthly", "yearly" ]
          },
          "appointment1Enabled" : {
            "type" : "boolean"
          },
          "appointment2Enabled" : {
            "type" : "boolean"
          },
          "appointment1Editable" : {
            "type" : "boolean"
          },
          "appointment2Editable" : {
            "type" : "boolean"
          }
        }
      },
      "DynamicTypeImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "attributes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AttributeImpl"
            }
          },
          "permissions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PermissionImpl"
            }
          },
          "name" : {
            "$ref" : "#/components/schemas/MultiLanguageName"
          },
          "key" : {
            "type" : "string"
          },
          "annotations" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/ParsedText"
            }
          }
        }
      },
      "ExternalSyncEntityImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "direction" : {
            "type" : "integer",
            "format" : "int32"
          },
          "context" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "raplaId" : {
            "type" : "string"
          },
          "externalSystem" : {
            "type" : "string"
          }
        }
      },
      "LinkReferenceHandler" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "linkType" : {
            "type" : "string"
          }
        }
      },
      "MultiLanguageName" : {
        "properties" : {
          "mapLocales" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
      "ParsedText" : {
        "properties" : {
          "formatString" : {
            "type" : "string"
          }
        }
      },
      "PermissionImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "pEnd" : {
            "type" : "string",
            "format" : "date-time"
          },
          "pStart" : {
            "type" : "string",
            "format" : "date-time"
          },
          "maxAdvance" : {
            "type" : "integer",
            "format" : "int32"
          },
          "minAdvance" : {
            "type" : "integer",
            "format" : "int32"
          },
          "accessLevel" : {
            "type" : "string",
            "enum" : [ "DENIED", "READ_TYPE", "CREATE", "READ_NO_ALLOCATION", "READ", "REQUEST", "ALLOCATE", "ALLOCATE_CONFLICTS", "EDIT", "ADMIN" ]
          }
        }
      },
      "PreferencePatch" : {
        "properties" : {
          "constants" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "configurations" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/RaplaConfiguration"
            }
          },
          "maps" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/RaplaMapImpl"
            }
          },
          "calendars" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/CalendarModelConfigurationImpl"
            }
          },
          "links" : {
            "$ref" : "#/components/schemas/LinkReferenceHandler"
          },
          "userId" : {
            "type" : "string"
          },
          "removedEntries" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PreferencesImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "map" : {
            "$ref" : "#/components/schemas/RaplaMapImpl"
          }
        }
      },
      "RaplaMapImpl" : {
        "properties" : {
          "constants" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "configurations" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/RaplaConfiguration"
            }
          },
          "calendars" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/CalendarModelConfigurationImpl"
            }
          },
          "links" : {
            "$ref" : "#/components/schemas/LinkReferenceHandler"
          }
        }
      },
...
      "ReservationImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "classification" : {
            "$ref" : "#/components/schemas/ClassificationImpl"
          },
          "appointments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AppointmentImpl"
            }
          },
          "permissions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PermissionImpl"
            }
          },
          "restrictions" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "annotations" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requestStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string",
              "enum" : [ "requested" ]
            }
          }
        }
      },
      "SerializableReferenceInfo" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "localname" : {
            "type" : "string"
          }
        }
      },
      "TimeInterval" : {
        "properties" : {
          "start" : {
            "type" : "string",
            "format" : "date-time"
          },
          "end" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "UpdateEvent" : {
        "properties" : {
          "categories" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CategoryImpl"
            }
          },
          "types" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DynamicTypeImpl"
            }
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserImpl"
            }
          },
          "preferencesPatches" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PreferencePatch"
            }
          },
          "preferences" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PreferencesImpl"
            }
          },
          "resources" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AllocatableImpl"
            }
          },
          "reservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReservationImpl"
            }
          },
          "conflicts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConflictImpl"
            }
          },
          "importExports" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExternalSyncEntityImpl"
            }
          },
          "removeSet" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SerializableReferenceInfo"
            },
            "uniqueItems" : true
          },
          "userId" : {
            "type" : "string"
          },
          "needResourcesRefresh" : {
            "type" : "boolean"
          },
          "forceAllocatableDeletesIgnoreDependencies" : {
            "type" : "boolean"
          },
          "invalidateInterval" : {
            "$ref" : "#/components/schemas/TimeInterval"
          },
          "lastValidated" : {
            "type" : "string"
          },
          "timezoneOffset" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "UserImpl" : {
        "properties" : {
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "id" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "admin" : {
            "type" : "boolean"
          },
          "authenticationSource" : {
            "type" : "string"
          },
          "lastChanged" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createDate" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ImpersonationResponse" : {
        "properties" : {
          "access_token" : {
            "type" : "string"
          },
          "token_type" : {
            "type" : "string"
          },
          "expires_in" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "oauth2" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "/oauth2/authorize",
            "tokenUrl" : "/oauth2/token",
            "scopes" : {
              "openid" : "OIDC subject identity",
              "profile" : "User profile claims"
            }
          }
        }
      },
      "bearerAuth" : {
        "type" : "http",
        "description" : "Paste a raw access token (e.g. from POST /oauth2/token grant_type=password). Manual fallback when OAuth2 isn't convenient.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}